Fix timestamp timezone
The code wrongly assumed that the value returned by
`datetime.fromtimestamp(ts)` returns a UTC datetime. However, the
returned datetime is a naive datetime in the local timezone. The code
would then previously replace the missing timezone with UTC.

For example, when running the code in CEST (UTC+2), this resulted in all
timestamps being shifted two hours in the future.

Instead, we can pass the UTC timezone directly as `tz` kwarg to
`datetime.fromtimestamp`, resulting in a non-naive datetime that has the
correct timezone attached.

An alternative approach would have been the following call:

    datetime.fromtimestamp(ts).astimezone()

...however using UTC based datetimes seems cleaner to me.